rename product snapshot to product list
